Split functions that require node into separate file (glob)
