+++ tsconfig.json
@@ -9,5 +9,5 @@
     "esModuleInterop": true,                   /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
     "typeRoots": ["./node_modules/@types", "./typings"]
   },
-  "exclude": ["node_modules", "**/*.test.ts", "typings"]
+  "exclude": ["node_modules", "**/*.test.ts", "typings", "playground"]
 }
